requirements.txt
curl_cffi==0.14.0
google-cloud-bigquery==3.40.1
google-generativeai==0.8.6
langdetect==1.0.9